more sheeva v pi (or debian?), also...python3 enters the chat

+++ scripts/standard/0_init.sh
@@ -86,7 +86,7 @@ install --verbose --mode=0755 --owner=0 --group=0 -D "${PFSRC}/dns-update.sh" "/
 install --verbose --mode=0755 --owner=0 --group=0 -D "${PFSRC}/ucarp-down-dhcpd.sh" "/usr/lib/untrustedhost/scripts/ucarp-down-dhcpd.sh"
 install --verbose --mode=0755 --owner=0 --group=0 -D "${PFSRC}/ucarp-up-dhcpd.sh" "/usr/lib/untrustedhost/scripts/ucarp-up-dhcpd.sh"
 
-[[ -e "${PFSRC}/${PACKER_BUILD_NAME}/fitstat" ]] && {
+if [[ -e "${PFSRC}/${PACKER_BUILD_NAME}/fitstat" ]] ; then
   install --verbose --mode=0755 --owner=0 --group=0 -D "${PFSRC}/${PACKER_BUILD_NAME}/fitstat" "/usr/lib/untrustedhost/scripts/fitstat"
   systemctl enable untrustedhost-fitstat
-}
+fi
